Create README.md
Dele is a fashion designer who needs an online platform in the form of a blog where
he can post his various categories of designs, and get feedback and thoughts from his
customers and blog visitors via comments. You are to build Restful APIs which
would be consumed by a client application.